requirements.txt
Django==1.7.4
Pillow==2.7.0
django-appconf==0.6
django-geoposition==0.2.2
psycopg2==2.5.4
six==1.9.0
wsgiref==0.1.2
